requirements.txt
astroid
Click
ecdsa
Flask
Flask-Cors
Flask-SQLAlchemy
flask-script
gunicorn
flask-migrate==2.4.0
future
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
pycryptodome<3.4.0
pylint
python-dotenv
python-jose-cryptodome
python-jose==3.0.1
psycopg2==2.8.4
six
SQLAlchemy==1.3.15
typed-ast
Werkzeug
wrapt